Move priority task into correct package
- Contains 4 levels of priority that can be assigned to tasks (LOW, MEDIUM, HIGH, URGENT)

It is important to note that this extension was implemented as a rolling form from week 1's optional `enums` task.
